Add files via upload
This application was created on 16-07-2021. It is an offline LMS which can be set up once by the university and the application can be distributed within the students.
It does not have any database compatibility right now and can handle 20 student accounts.
It is in its initial stage and may not be that capable now but with regular commits , I assure a robust and fully functional application soon. With new features coming soon I assure you all to a much better and useful utility software.